MAINT: calculate_velocity_texture moved to pyart.retrieve
