+++ pyart/filters/__init__.py
@@ -21,6 +21,5 @@ Filtering radar data
 
 from .gatefilter import GateFilter, moment_based_gate_filter
 from .gatefilter import moment_and_texture_based_gate_filter
-from .gatefilter import calculate_velocity_texture
 
 __all__ = [s for s in dir() if not s.startswith('_')]
