+++ pyart/retrieve/__init__.py
@@ -42,6 +42,7 @@ from .echo_class import steiner_conv_strat, hydroclass_semisupervised
 from .echo_class import get_freq_band
 from .gate_id import map_profile_to_gates, fetch_radar_time_profile
 from .simple_moment_calculations import calculate_snr_from_reflectivity
+from .simple_moment_calculations import calculate_velocity_texture
 from .simple_moment_calculations import compute_snr, compute_l, compute_cdr
 from .simple_moment_calculations import compute_noisedBZ
 from .advection import grid_displacement_pc, grid_shift
