+++ pyart/retrieve/simple_moment_calculations.py
@@ -17,8 +17,10 @@ Simple moment calculations.
 
 import numpy as np
 
+from scipy import ndimage
 from ..config import get_metadata, get_field_name, get_fillvalue
 from ..core.transforms import antenna_to_cartesian
+from ..util import angular_texture_2d
 
 
 def calculate_snr_from_reflectivity(
@@ -251,3 +253,68 @@ def compute_cdr(radar, rhohv_field=None, zdr_field=None, cdr_field=None):
     cdr['data'] = cdr_data
 
     return cdr
+
+
+def calculate_velocity_texture(radar, vel_field=None, wind_size=4, nyq=None,
+                               check_nyq_uniform=True):
+    """
+    Derive the texture of the velocity field
+
+    Parameters
+    ----------
+    radar: Radar
+        Radar object from which velocity texture field will be made.
+    vel_field_name : str
+        Name of the velocity field. A value of None will force Py-ART to
+        automatically determine the name of the velocity field.
+    wind_size : int
+        The size of the window to calculate texture from. The window is
+        defined to be a square of size wind_size by wind_size.
+    nyq : float
+        The nyquist velocity of the radar. A value of None will force Py-ART
+        to try and determine this automatically.
+    check_nyquist_uniform : bool, optional
+        True to check if the Nyquist velocities are uniform for all rays
+        within a sweep, False will skip this check. This parameter is ignored
+        when the nyq parameter is not None.
+
+    Returns
+    -------
+    vel_dict: dict
+        A dictionary containing the field entries for the radial velocity
+        texture.
+
+    """
+
+    # Parse names of velocity field
+    if vel_field is None:
+        vel_field = get_field_name('velocity')
+
+    # Allocate memory for texture field
+    vel_texture = np.zeros(radar.fields[vel_field]['data'].shape)
+
+    # If an array of nyquist velocities is derived, use different
+    # nyquist velocites for each sweep in texture calculation according to
+    # the nyquist velocity in each sweep.
+
+    if(nyq is None):
+        # Find nyquist velocity if not specified
+        nyq = [radar.get_nyquist_vel(i, check_nyq_uniform) for i in
+               range(radar.nsweeps)]
+        for i in range(0, radar.nsweeps):
+            start_ray, end_ray = radar.get_start_end(i)
+            inds = range(start_ray, end_ray)
+            vel_sweep = radar.fields[vel_field]['data'][inds]
+            vel_texture[inds] = angular_texture_2d(
+                vel_sweep, wind_size, nyq[i])
+    else:
+        vel_texture = angular_texture_2d(
+            radar.fields[vel_field]['data'], wind_size, nyq)
+    vel_texture_field = get_metadata('velocity')
+    vel_texture_field['long_name'] = 'Doppler velocity texture'
+    vel_texture_field['standard_name'] = ('texture_of_radial_velocity' +
+                                          '_of_scatters_away_from_instrument')
+    vel_texture_field['data'] = ndimage.filters.median_filter(vel_texture,
+                                                              size=(wind_size,
+                                                                    wind_size))
+    return vel_texture_field
