+++ pyart/retrieve/tests/test_simple_moment_calculations.py
@@ -15,3 +15,25 @@ def test_calculate_snr_from_reflectivity():
     test_radar.add_field('reflectivity', foo_field)
     snr = pyart.retrieve.calculate_snr_from_reflectivity(test_radar, toa=500)
     assert snr['data'].mean() < 1e-6
+
+
+
+def test_calculate_velocity_texture():
+    radar = pyart.testing.make_empty_ppi_radar(10, 36, 1)
+
+    # a zero field
+    fdata = np.tile(np.arange(10.), 36).reshape(36, 10)
+    fdata3 = np.zeros(fdata.shape)
+    radar.add_field('zero_field', {'data': fdata3})
+
+    vel_field = 'zero_field'
+    texture_field = pyart.retrieve.calculate_velocity_texture(
+        radar, vel_field, wind_size=2, nyq=10)
+    assert np.all(texture_field['data'] == 0)
+    texture_field = pyart.retrieve.calculate_velocity_texture(
+        radar, vel_field, wind_size=3, nyq=10)
+    assert np.all(texture_field['data'] == 0)
+    texture_field = pyart.retrieve.calculate_velocity_texture(
+        radar, vel_field, wind_size=4, nyq=10)
+    assert np.all(texture_field['data'] == 0)
+
